Moved selector logic to a dedicated crate.
